+++ app/assets/javascript/ahoy_captain/controllers/interval_controller.js
@@ -1,0 +1,10 @@
+import {Controller} from "@hotwired/stimulus"
+
+export default class extends Controller {
+  handleChange(event) {
+    const url = new URL(event.target.form.action);
+    const interval = event.target.value;
+    url.searchParams.set("interval", interval)
+    event.target.closest('turbo-frame').src = url.href
+  }
+}