+++ app/controllers/ahoy_captain/stats/bounce_rates_controller.rb
@@ -4,7 +4,7 @@ module AhoyCaptain
       # @todo: this is lazy
       def index
         @stats = AhoyCaptain::Stats::BounceRatesQuery.call(params)
-        @stats = @stats.group_by_day("daily_bounce_rate.date").average("bounce_rate")
+        @stats = @stats.group_by_period(selected_interval, "daily_bounce_rate.date").average("bounce_rate")
       end
     end
   end