+++ app/controllers/ahoy_captain/stats/total_pageviews_controller.rb
@@ -2,7 +2,7 @@ module AhoyCaptain
   module Stats
     class TotalPageviewsController < BaseController
       def index
-        @stats = AhoyCaptain::Stats::TotalPageviewsQuery.call(params).group_by_day(:time).count
+        @stats = AhoyCaptain::Stats::TotalPageviewsQuery.call(params).group_by_period(selected_interval, :time).count
       end
     end
   end