+++ app/controllers/ahoy_captain/stats/total_visits_controller.rb
@@ -2,7 +2,7 @@ module AhoyCaptain
   module Stats
     class TotalVisitsController < BaseController
       def index
-        @stats = AhoyCaptain::Stats::TotalVisitorsQuery.call(params).group_by_day(:started_at).count
+        @stats = AhoyCaptain::Stats::TotalVisitorsQuery.call(params).group_by_period(selected_interval, :started_at).count
       end
     end
   end