+++ app/controllers/ahoy_captain/stats/unique_visitors_controller.rb
@@ -2,7 +2,7 @@ module AhoyCaptain
   module Stats
     class UniqueVisitorsController < BaseController
       def index
-        @stats = AhoyCaptain::Stats::UniqueVisitorsQuery.call(params).group_by_day(:started_at).count
+        @stats = AhoyCaptain::Stats::UniqueVisitorsQuery.call(params).group_by_period(selected_interval, :started_at).count
       end
     end
   end