+++ app/controllers/ahoy_captain/stats/views_per_visits_controller.rb
@@ -5,7 +5,7 @@ module AhoyCaptain
 
       # @todo: make me a window func
       def index
-        @stats = AhoyCaptain::Stats::ViewsPerVisitQuery.call(params).group_by_day('views_per_visit_table.started_at').average(:views_per_visit)
+        @stats = AhoyCaptain::Stats::ViewsPerVisitQuery.call(params).group_by_period(selected_interval, 'views_per_visit_table.started_at').average(:views_per_visit)
         (range[0].to_date..range[1].to_date).to_a.each do |date|
           unless @stats.key?(date.to_date)
             @stats[date.to_date] = 2