+++ app/controllers/ahoy_captain/stats/visit_durations_controller.rb
@@ -2,7 +2,7 @@ module AhoyCaptain
   module Stats
     class VisitDurationsController < BaseController
       def index
-        @stats = AhoyCaptain::Stats::VisitDurationQuery.call(params).group('started_at').average(:duration)
+        @stats = AhoyCaptain::Stats::VisitDurationQuery.call(params).group_by_period(selected_interval, 'started_at').average(:duration)
       end
     end
   end